requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alphashape==1.3.1
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
beautifulsoup4==4.14.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
click-log==0.4.0
colorama==0.4.6
Crawl4AI==0.7.4
cryptography==46.0.1
distro==1.9.0
faiss-cpu==1.12.0
fake-http-header==0.3.5
fake-useragent==2.2.0
fastapi==0.117.1
fastuuid==0.13.5
ffmpy==0.6.1
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gradio==5.47.2
gradio_client==1.13.3
greenlet==3.2.4
groovy==0.1.2
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.1
humanize==4.13.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langdetect==1.0.9
lark==1.3.0
litellm==1.77.4
lxml==5.4.0
madoka @ file:///D:/bld/madoka_1758567393280/work
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
networkx==3.5
nltk==3.9.1
numpy==2.3.3
openai==1.109.1
orjson==3.11.3
packaging==25.0
pandas==2.3.2
patchright==1.55.2
pillow==11.3.0
playwright==1.55.0
pondpond==1.4.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
rank-bm25==0.2.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
rtree==1.4.1
ruff==0.13.2
safehttpx==0.1.6
safetensors==0.6.2
scipy==1.16.2
semantic-version==2.10.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.8
starlette==0.48.0
sympy==1.14.0
tf-playwright-stealth==1.2.0
tiktoken==0.11.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.8.0+cu129
tqdm==4.67.1
transformers==4.56.2
trimesh==4.8.3
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
